+++ src/utilities/mod.rs
@@ -2,5 +2,4 @@
 //! as well as decode keys from different formats.
 
 pub(crate) mod base64;
-pub mod key_decoder;
 pub mod key_generation;