+++ tests/diesel-postgres/main.rs
@@ -7,14 +7,13 @@ use encrypted_message::{
     EncryptedMessage,
     strategy::{Randomized, Deterministic},
     key_config::Secret,
-    utilities::key_decoder::{KeyDecoder as _, HexKeyDecoder},
 };
 
 #[derive(Debug, Default)]
 struct KeyConfig;
 impl encrypted_message::KeyConfig for KeyConfig {
     fn keys(&self) -> Vec<Secret<[u8; 32]>> {
-        HexKeyDecoder::decode_keys(vec![String::from("75754f7866705767526749456f33644972646f30686e484a484631686e747657").into()])
+        vec![(*b"uuOxfpWgRgIEo3dIrdo0hnHJHF1hntvW").into()]
     }
 }